Performance, conversation and extractor fixes
replaced some memory allocation heavy parts, fixed the bool handling in
the extractor (which caused it to skip further properties), changed the
resourceID in the conversation packed to the relevanceID (which also
makes the client window display the partner's name correctly). Btw, some
conversations are still not extracted correctly. (the one guard i tested
had an additional dialog from a different npc-type). I think the
additional topics option is there for quests related dialogs not to
include more conversation topics (which was the case with this guard).